+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/Transition.hs
@@ -11,7 +11,6 @@ import Cardano.Ledger.Allegra.Translation ()
 import Cardano.Ledger.Genesis (NoGenesis (..))
 import Cardano.Ledger.Shelley
 import Cardano.Ledger.Shelley.Transition
-import Data.Aeson (FromJSON (..), ToJSON (..))
 import Lens.Micro
 import NoThunks.Class (NoThunks (..))
 
@@ -19,7 +18,7 @@ instance EraTransition AllegraEra where
   newtype TransitionConfig AllegraEra = AllegraTransitionConfig
     { atcShelleyTransitionConfig :: TransitionConfig ShelleyEra
     }
-    deriving (Show, Eq, NoThunks, ToJSON, FromJSON)
+    deriving (Show, Eq, NoThunks)
 
   mkTransitionConfig NoGenesis = AllegraTransitionConfig
 
@@ -28,4 +27,4 @@ instance EraTransition AllegraEra where
   tcPreviousEraConfigL =
     lens atcShelleyTransitionConfig (\atc pc -> atc {atcShelleyTransitionConfig = pc})
 
-  tcTranslationContextL = lens (const NoGenesis) (const . id)
+  tcTranslationContextL = lens (const NoGenesis) const