+++ eras/dijkstra/src/Cardano/Ledger/Dijkstra/Genesis.hs
@@ -1,15 +1,16 @@
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Dijkstra.Genesis (
   DijkstraGenesis (..),
-  toDijkstraGenesisPairs,
 ) where
 
+import Cardano.Ledger.BaseTypes (KeyValuePairs (..), ToKeyValuePairs (..))
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra)
 import Cardano.Ledger.Genesis (EraGenesis (..))
-import Data.Aeson (FromJSON, ToJSON)
+import Data.Aeson (FromJSON (..), ToJSON, withObject)
 import GHC.Generics
 import NoThunks.Class (NoThunks)
 
@@ -17,10 +18,10 @@ import NoThunks.Class (NoThunks)
 -- in the Dijkstra era
 data DijkstraGenesis = DijkstraGenesis
   deriving (Eq, Show, Generic)
+  deriving (ToJSON) via KeyValuePairs DijkstraGenesis
 
-instance ToJSON DijkstraGenesis
-
-instance FromJSON DijkstraGenesis
+instance FromJSON DijkstraGenesis where
+  parseJSON = withObject "DijkstraGenesis" $ \_ -> pure DijkstraGenesis
 
 instance NoThunks DijkstraGenesis
 
@@ -28,5 +29,5 @@ instance EraGenesis DijkstraEra where
   type Genesis DijkstraEra = DijkstraGenesis
 
 -- TODO: Implement this and use for ToJSON instance
-toDijkstraGenesisPairs :: DijkstraGenesis -> [a]
-toDijkstraGenesisPairs _ = []
+instance ToKeyValuePairs DijkstraGenesis where
+  toKeyValuePairs DijkstraGenesis = []