+++ eras/dijkstra/src/Cardano/Ledger/Dijkstra/Transition.hs
@@ -10,29 +10,17 @@
 
 module Cardano.Ledger.Dijkstra.Transition (
   TransitionConfig (..),
-  toDijkstraTransitionConfigPairs,
 ) where
 
 import Cardano.Ledger.Conway
 import Cardano.Ledger.Conway.Transition (
   ConwayEraTransition,
   registerDRepsThenDelegs,
-  toConwayTransitionConfigPairs,
  )
 import Cardano.Ledger.Dijkstra.Era
 import Cardano.Ledger.Dijkstra.Genesis
 import Cardano.Ledger.Dijkstra.Translation ()
 import Cardano.Ledger.Shelley.Transition
-import Data.Aeson (
-  FromJSON (..),
-  KeyValue (..),
-  ToJSON (..),
-  Value (..),
-  object,
-  pairs,
-  withObject,
-  (.:),
- )
 import GHC.Generics
 import Lens.Micro
 import NoThunks.Class (NoThunks (..))
@@ -59,20 +47,3 @@ instance EraTransition DijkstraEra where
 instance ConwayEraTransition DijkstraEra
 
 instance NoThunks (TransitionConfig DijkstraEra)
-
-instance ToJSON (TransitionConfig DijkstraEra) where
-  toJSON = object . toDijkstraTransitionConfigPairs
-  toEncoding = pairs . mconcat . toDijkstraTransitionConfigPairs
-
-toDijkstraTransitionConfigPairs :: KeyValue e a => TransitionConfig DijkstraEra -> [a]
-toDijkstraTransitionConfigPairs dijkstraConfig =
-  toConwayTransitionConfigPairs conwayConfig
-    ++ ["dijkstra" .= object (toDijkstraGenesisPairs (dijkstraConfig ^. tcTranslationContextL))]
-  where
-    conwayConfig = dijkstraConfig ^. tcPreviousEraConfigL
-
-instance FromJSON (TransitionConfig DijkstraEra) where
-  parseJSON = withObject "DijkstraTransitionConfig" $ \o -> do
-    pc <- parseJSON (Object o)
-    ag <- o .: "dijkstra"
-    pure $ mkTransitionConfig pc ag