+++ eras/dijkstra/testlib/Test/Cardano/Ledger/Dijkstra/Arbitrary.hs
@@ -5,6 +5,8 @@ module Test.Cardano.Ledger.Dijkstra.Arbitrary () where
 
 import Cardano.Ledger.Dijkstra (DijkstraEra)
 import Cardano.Ledger.Dijkstra.Core (EraTxBody (..))
+import Cardano.Ledger.Dijkstra.Genesis (DijkstraGenesis (..))
+import Cardano.Ledger.Dijkstra.Transition (TransitionConfig (..))
 import Cardano.Ledger.Dijkstra.TxBody (TxBody (..))
 import Test.Cardano.Ledger.Common (Arbitrary (..), scale)
 import Test.Cardano.Ledger.Conway.Arbitrary ()
@@ -31,3 +33,9 @@ instance Arbitrary (TxBody DijkstraEra) where
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
+
+instance Arbitrary DijkstraGenesis where
+  arbitrary = pure DijkstraGenesis
+
+instance Arbitrary (TransitionConfig DijkstraEra) where
+  arbitrary = DijkstraTransitionConfig <$> arbitrary <*> arbitrary