+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Translation.hs
@@ -1,8 +1,9 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
@@ -12,6 +13,7 @@ module Cardano.Ledger.Shelley.Translation (
   toFromByronTranslationContext,
 ) where
 
+import Cardano.Ledger.BaseTypes (KeyValuePairs (..), ToKeyValuePairs (..))
 import Cardano.Ledger.Binary (
   DecCBOR (..),
   EncCBOR (..),
@@ -26,6 +28,7 @@ import Cardano.Ledger.Core (PParams, TranslationContext, emptyPParams)
 import Cardano.Ledger.Keys
 import Cardano.Ledger.Shelley.Era (ShelleyEra)
 import Cardano.Ledger.Shelley.Genesis (ShelleyGenesis (..))
+import Data.Aeson (FromJSON (..), ToJSON (..), withObject, (.:), (.=))
 import Data.Map (Map)
 import qualified Data.Map as Map
 import GHC.Generics (Generic)
@@ -39,6 +42,7 @@ data FromByronTranslationContext = FromByronTranslationContext
   , fbtcMaxLovelaceSupply :: !Word64
   }
   deriving (Eq, Show, Generic)
+  deriving (ToJSON) via KeyValuePairs FromByronTranslationContext
 
 instance ToCBOR FromByronTranslationContext where
   toCBOR x@(FromByronTranslationContext _ _ _) =
@@ -62,6 +66,21 @@ instance FromCBOR FromByronTranslationContext where
 instance EncCBOR FromByronTranslationContext
 
 instance DecCBOR FromByronTranslationContext
+
+instance FromJSON FromByronTranslationContext where
+  parseJSON = withObject "FromByronTranslationContext" $ \o -> do
+    fbtcGenDelegs <- o .: "genDelegs"
+    fbtcProtocolParams <- o .: "protocolParams"
+    fbtcMaxLovelaceSupply <- o .: "maxLovelaceSupply"
+    pure FromByronTranslationContext {..}
+
+instance ToKeyValuePairs FromByronTranslationContext where
+  toKeyValuePairs fbtc@(FromByronTranslationContext _ _ _) =
+    let FromByronTranslationContext {..} = fbtc
+     in [ "genDelegs" .= fbtcGenDelegs
+        , "protocolParams" .= fbtcProtocolParams
+        , "maxLovelaceSupply" .= fbtcMaxLovelaceSupply
+        ]
 
 -- | Trivial FromByronTranslationContext value, for use in cases where we do not need
 -- to translate from Byron to Shelley.