+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Era.hs
@@ -10,6 +10,7 @@ module Test.Cardano.Ledger.Shelley.Era (
 import Cardano.Ledger.Shelley
 import Cardano.Ledger.Shelley.LedgerState
 import Cardano.Ledger.Shelley.Scripts
+import Cardano.Ledger.Shelley.Transition
 import Cardano.Ledger.State
 import Data.Default
 import Test.Cardano.Ledger.Common
@@ -20,6 +21,8 @@ import Test.Cardano.Ledger.Shelley.TreeDiff ()
 class
   ( EraTest era
   , ShelleyEraScript era
+  , EraTransition era
+  , Arbitrary (TransitionConfig era)
   , Eq (StashedAVVMAddresses era)
   , Show (StashedAVVMAddresses era)
   , ToExpr (StashedAVVMAddresses era)