+++ libs/cardano-ledger-core/src/Cardano/Ledger/Genesis.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE DerivingVia #-}
+{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableInstances #-}
@@ -8,6 +10,7 @@ module Cardano.Ledger.Genesis (
   NoGenesis (..),
 ) where
 
+import Cardano.Ledger.BaseTypes (KeyValuePairs (..), ToKeyValuePairs (..))
 import Cardano.Ledger.Binary (
   DecCBOR (..),
   EncCBOR (..),
@@ -15,6 +18,13 @@ import Cardano.Ledger.Binary (
   ToCBOR (..),
  )
 import Cardano.Ledger.Core.Era (Era)
+import Control.Monad (unless)
+import Data.Aeson (
+  FromJSON (..),
+  ToJSON,
+  Value (..),
+ )
+import qualified Data.Aeson.KeyMap as KV
 import Data.Kind (Type)
 
 class Era era => EraGenesis era where
@@ -23,6 +33,7 @@ class Era era => EraGenesis era where
 
 data NoGenesis era = NoGenesis
   deriving (Eq, Show)
+  deriving (ToJSON) via KeyValuePairs (NoGenesis era)
 
 instance Era era => ToCBOR (NoGenesis era) where
   toCBOR _ = toCBOR ()
@@ -33,3 +44,14 @@ instance Era era => FromCBOR (NoGenesis era) where
 instance Era era => EncCBOR (NoGenesis era)
 
 instance Era era => DecCBOR (NoGenesis era)
+
+instance ToKeyValuePairs (NoGenesis era) where
+  toKeyValuePairs _ = []
+
+instance FromJSON (NoGenesis era) where
+  parseJSON = \case
+    Null -> pure NoGenesis
+    Object o -> do
+      unless (KV.null o) $ fail "NoGenesis cannot have any fields"
+      pure NoGenesis
+    _ -> fail "Unexpected value type for NoGenesis"