+++ eras/shelley/impl/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 1.17.0.0
 
-* Deprecated `toShelleyGenesisPairs`
+* Add `ToJSON` and `FromJSON` instances for `FromByronTranslationContext`
+* Deprecated `toShelleyGenesisPairs` and `toShelleyTransitionConfigPairs`
 * Removed `toShelleyGenesisPairs`
 * Remove `ShelleyTxRaw`, `MkShelleyTx`, `segWitTx`, `unsafeConstructTxWithBytes`
 * Added `Generic` instances for:
@@ -42,6 +43,7 @@
 
 ### `testlib`
 
+* Added `Arbitrary` instance for `TransitionConfig ShelleyEra`
 * Rename `poolParams` to `freshPoolParams`
 * Added `ToExpr` instances for:
   * `ShelleyScriptsNeeded`